Refactor based on PR review
